requirements.txt
openai-agents
openai
python-dotenv
faiss-cpu
rich
asyncio
aiohttp
redis>=5.0.0
pydantic-settings>=2.2.1
uvicorn[standard]
fastapi-limiter>=0.1.6
pydantic_settings
pytest
mem0ai
PyYAML>=6.0
